Move Freelex code into a module so that we can add Signbank equivalents alongside
