+++ app/models/freelex/sign_parser.rb
@@ -1,0 +1,115 @@
+# frozen_string_literal: true
+
+module Freelex
+  class SignParser
+    VIDEO_EXAMPLES_TOTAL = 4
+
+    BOOLEAN_MAP = {
+      contains_numbers: 'number_incorp',
+      is_fingerspelling: 'fingerspelling',
+      is_directional: 'directional',
+      is_locatable: 'locatable',
+      one_or_two_handed: 'one_or_two_hand'
+    }.freeze
+
+    TAG_MAP = {
+      age_groups: 'VARIATIONAGE',
+      gender_groups: 'VARIATIONGENDER',
+      hint: 'hint',
+      usage: 'usage',
+      usage_notes: 'essay',
+      related_to: 'RELATEDTO',
+      gloss_main: 'glossmain',
+      gloss_secondary: 'glosssecondary',
+      gloss_minor: 'glossminor',
+      gloss_maori: 'glossmaori',
+      drawing: 'ASSET picture',
+      handshape: 'handshape',
+      location_name: 'location',
+      word_classes: 'SECONDARYWORDCLASS',
+      inflection: 'INFLECTION'
+    }.freeze
+
+    # @param data [Nokogiri::XML::Element]
+    def initialize(data)
+      @data = data
+    end
+
+    def build_sign
+      @sign = Sign.new
+      return sign if @data.nil?
+
+      @sign.id = @data.value_for_tag('headwordid')
+
+      parse_video
+      parse_tags
+      parse_booleans
+      parse_examples
+      build_inflections
+
+      @sign
+    end
+
+    def build_inflections
+      @sign.inflection_plural = @sign.inflection.match('plural')
+      @sign.inflection_temporal = @sign.inflection.match('temporal')
+      @sign.inflection_manner_and_degree = @sign.inflection.match('manner')
+    end
+
+    def parse_video
+      @sign.video = "#{ASSET_URL}#{@data.value_for_tag('ASSET glossmain')}"
+      glossmain_slow = @data.value_for_tag('ASSET glossmain_slow')
+      @sign.video_slow = "#{ASSET_URL}#{@data.value_for_tag('ASSET glossmain_slow')}" if glossmain_slow.present?
+    end
+
+    def parse_tags
+      TAG_MAP.each do |key, tag|
+        value = @data.value_for_tag(tag)
+        @sign.send("#{key}=", value)
+      end
+    end
+
+    def parse_booleans
+      BOOLEAN_MAP.each do |key, tag|
+        value = @data.value_for_tag(tag).to_bool
+        @sign.send("#{key}=", value)
+      end
+    end
+
+    def parse_examples
+      # examples
+      @sign.examples = []
+      VIDEO_EXAMPLES_TOTAL.times do |i|
+        next if @data.value_for_tag("ASSET finalexample#{i}").blank?
+
+        video_slow = if @data.value_for_tag("ASSET finalexample#{i}_slow").present?
+                       "#{ASSET_URL}#{@data.value_for_tag("ASSET finalexample#{i}_slow")}"
+                     end
+
+        @sign.examples << { transcription: parse_transcription("videoexample#{i}"),
+                            translation: @data.value_for_tag("videoexample#{i}translation"),
+                            video: "#{ASSET_URL}#{@data.value_for_tag("ASSET finalexample#{i}")}",
+                            video_slow: video_slow }
+      end
+    end
+
+    def parse_transcription(tag)
+      transcription = []
+      @data.css(tag).children.each do |item|
+        if item.is_a?(Nokogiri::XML::Text)
+          transcription += item.content.split(' ')
+        else
+          transcription << { id: item['id'], gloss: item.children.first.content }
+        end
+      end
+      transcription
+    end
+    # Extend Nokogiri with helper method for fetching value
+    Nokogiri::XML::Element.class_eval do
+      def value_for_tag(tag_name)
+        tag = css(tag_name).first
+        tag.is_a?(Nokogiri::XML::Node) ? tag.content : ''
+      end
+    end
+  end
+end
