+++ spec/models/freelex/sign_menu_spec.rb
@@ -1,0 +1,16 @@
+# frozen_string_literal: true
+
+require 'rails_helper'
+
+module Freelex
+  describe 'SignMenu' do
+    it 'must not throw errors' do
+      # call the classmethods
+      SignMenu.usage_tags
+      SignMenu.topic_tags
+      SignMenu.locations
+      SignMenu.location_groups
+      # TODO: check returned result
+    end
+  end
+end
